Converting int32 money digits to int64 and closing the issues related to installation.

+++ Abbey Trading Store/UI/Advanced/Screen_forms/Overview.cs
@@ -59,10 +59,10 @@ namespace Abbey_Trading_Store.UI.Advanced.Screen_forms
                 {
                     
                     chart1.Series["Sales"].Points.AddXY(dr[2].ToString(), Convert.ToInt32(dr[1]));
-                    chart1.Series["Profit Accummulation"].Points.AddXY(dr[2].ToString(), Convert.ToInt32(dr[0]));
+                    chart1.Series["Profit Accummulation"].Points.AddXY(dr[2].ToString(), Convert.ToInt64(dr[0]));
 
-                    label6.Text = "Shs. " + Convert.ToInt32(dr[1]).ToString("N0");
-                    profitlbl.Text = "Shs. " + Convert.ToInt32(dr[0]).ToString("N0");
+                    label6.Text = "Shs. " + Convert.ToInt64(dr[1]).ToString("N0");
+                    profitlbl.Text = "Shs. " + Convert.ToInt64(dr[0]).ToString("N0");
 
                     circularProgressBar1.Value = Convert.ToInt32(dr[1]) / 1000000000;
                     Profits.Value = Convert.ToInt32(dr[0]) / 1000000000;
@@ -75,7 +75,7 @@ namespace Abbey_Trading_Store.UI.Advanced.Screen_forms
 
                 foreach(DataRow dr in dt_4.Rows)
                 {
-                    chart1.Series["Expenditure"].Points.AddXY(dr[1].ToString(), Convert.ToInt32(dr[0]));
+                    chart1.Series["Expenditure"].Points.AddXY(dr[1].ToString(), Convert.ToInt64(dr[0]));
 
                 }
 
@@ -120,17 +120,17 @@ namespace Abbey_Trading_Store.UI.Advanced.Screen_forms
                 DataTable dt_3 = DAL.DAL_Properties.Env.mode == 1 ? (user.TransactionsOverview_2()) : (user.TransactionsOverview("Customer"));
                 foreach (DataRow dr in dt_3.Rows)
                 {
-                    chart1.Series["Transactions"].Points.AddXY(dr[2].ToString(), Convert.ToInt32(dr[1]));
-                    chart1.Series["Profit Accummulation"].Points.AddXY(dr[2].ToString(), Convert.ToInt32(dr[0]));
+                    chart1.Series["Transactions"].Points.AddXY(dr[2].ToString(), Convert.ToInt64(dr[1]));
+                    chart1.Series["Profit Accummulation"].Points.AddXY(dr[2].ToString(), Convert.ToInt64(dr[0]));
 
-                    label6.Text = "Shs. " + Convert.ToInt32(dr[1]).ToString("N0");
-                    profitlbl.Text = "Shs. " + Convert.ToInt32(dr[0]).ToString("N0");
+                    label6.Text = "Shs. " + Convert.ToInt64(dr[1]).ToString("N0");
+                    profitlbl.Text = "Shs. " + Convert.ToInt64(dr[0]).ToString("N0");
 
                     circularProgressBar1.Value = Convert.ToInt32(dr[1]) / 1000000000;
                     Profits.Value = Convert.ToInt32(dr[0]) / 1000000000;
 
-                    circularProgressBar1.Text = Convert.ToInt32(dr[1]) / 1000000000 * 100 + "%";
-                    Profits.Text = Convert.ToInt32(dr[0]) / 1000000000 * 100 + "%";
+                    circularProgressBar1.Text = Convert.ToInt64(dr[1]) / 1000000000 * 100 + "%";
+                    Profits.Text = Convert.ToInt64(dr[0]) / 1000000000 * 100 + "%";
